pyproject.toml
[project]
name = "ha-memory"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel Sendzik",email = "daniel@sendzik.eu"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "<4.0,>=3.13"
dependencies = [
    "fastmcp (>=2.11.3,<3.0.0)",
    "aiohttp (>=3.8.0,<4.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "matrix-nio (>=0.25.2,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
